Moved fork_and_pipe to exec_pipes
